Testing label detection. --ChufanLai

+++ apis/annotation/libs/objattrs.py
@@ -234,7 +234,7 @@ class ObjAttrs:
             return contour_list
         img = self.img
         # Set up the mask image
-        mask_img = np.zeros(img.shape[:2])
+        mask_img = np.zeros(img.shape[:2], dtype=np.uint8)
         bbox_mask = np.zeros(img.shape[:2])
         bbox_poly = np.array([\
             [bbox["x"], bbox["y"]],\
